+++ source/common/src/msg/tmsg.c
@@ -13170,57 +13170,6 @@ _exit:
   return code;
 }
 
-int32_t tSerializeSMqSeekReq(void *buf, int32_t bufLen, SMqSeekReq *pReq) {
-  int32_t code = 0;
-  int32_t lino;
-  int32_t headLen = sizeof(SMsgHead);
-  if (buf != NULL) {
-    buf = (char *)buf + headLen;
-    bufLen -= headLen;
-  }
-  SEncoder encoder = {0};
-  tEncoderInit(&encoder, buf, bufLen);
-  TAOS_CHECK_EXIT(tStartEncode(&encoder));
-  TAOS_CHECK_EXIT(tEncodeI64(&encoder, pReq->consumerId));
-  TAOS_CHECK_EXIT(tEncodeCStr(&encoder, pReq->subKey));
-  tEndEncode(&encoder);
-
-_exit:
-  if (code) {
-    tEncoderClear(&encoder);
-    return code;
-  } else {
-    int32_t tlen = encoder.pos;
-    tEncoderClear(&encoder);
-
-    if (buf != NULL) {
-      SMsgHead *pHead = (SMsgHead *)((char *)buf - headLen);
-      pHead->vgId = htonl(pReq->head.vgId);
-      pHead->contLen = htonl(tlen + headLen);
-    }
-
-    return tlen + headLen;
-  }
-}
-
-int32_t tDeserializeSMqSeekReq(void *buf, int32_t bufLen, SMqSeekReq *pReq) {
-  int32_t  code = 0;
-  int32_t  lino;
-  int32_t  headLen = sizeof(SMsgHead);
-  SDecoder decoder = {0};
-  tDecoderInit(&decoder, (char *)buf + headLen, bufLen - headLen);
-
-  TAOS_CHECK_EXIT(tStartDecode(&decoder));
-  TAOS_CHECK_EXIT(tDecodeI64(&decoder, &pReq->consumerId));
-  TAOS_CHECK_EXIT(tDecodeCStrTo(&decoder, pReq->subKey));
-
-  tEndDecode(&decoder);
-
-_exit:
-  tDecoderClear(&decoder);
-  return code;
-}
-
 int32_t tSerializeSDownstreamSourceNode(SEncoder *pEncoder, SDownstreamSourceNode *pSource) {
   int32_t code = 0;
   int32_t lino;
