+++ source/dnode/vnode/src/vnd/vnodeApi.c
@@ -95,7 +95,7 @@ void initMetadataAPI(SStoreMeta* pMeta) {
 
   pMeta->extractTagVal = (const void* (*)(const void*, int16_t, STagVal*))metaGetTableTagVal;
   pMeta->getTableTags = metaGetTableTags;
-  pMeta->getTableTagsByUid = metaGetTableTagsByUids;
+  pMeta->getTableTagsByUidVersion = metaGetTableTagsByUidsVersion;
 
   pMeta->getTableUidByName = metaGetTableUidByName;
   pMeta->getTableTypeSuidByName = metaGetTableTypeSuidByName;
